fcp-microservices/services/workbench/db_dashboard_api/requirements.txt
uvicorn[standard]
fastapi
magnum
psycopg2-binary
mangum
requests
python-dotenv